Refactor worker & Add model runner
